Chore: refactor create and editModule + few style changes of permissions in module creation
