+++ packages/web-main/src/pages/ModuleDefinitions/style.ts
@@ -12,13 +12,30 @@ export const PermissionCard = styled.li`
   width: 100%;
   display: flex;
   align-items: center;
+  flex-direction: column;
 `;
+
 export const ContentContainer = styled.div`
   width: 100%;
   display: flex;
   flex-direction: column;
   padding: ${({ theme }) => theme.spacing[1]};
   gap: ${({ theme }) => theme.spacing[4]};
+`;
+
+export const Title = styled.div`
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  margin-top: ${({ theme }) => theme.spacing[2]};
+  margin-bottom: ${({ theme }) => theme.spacing['0_5']};
+
+  .inner {
+    display: flex;
+    gap: ${({ theme }) => theme.spacing['0_5']};
+  }
 `;
 
 export const Fields = styled.div`
@@ -38,10 +55,4 @@ export const Flex = styled.div`
   & > *:first-child {
     flex-grow: 1;
   }
-`;
-
-export const Column = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: ${({ theme }) => theme.spacing[1]};
 `;
